# Conflicts:
#	server/controllers/posts.js

+++ server/controllers/posts.js
@@ -7,9 +7,11 @@ const {
   getAllPosts
 } = require('../services/post-services.js');
 
-const { removeLikesFromPost } = require('../services/like-services');
+const { removeLikesFromPost, removeLikesFromComment } = require('../services/like-services');
 
-const { removeAllPostComments } = require('../services/comment-services');
+const {
+  getCommentsOfPost, removeAllPostComments
+} = require('../services/comment-services');
 
 const serverResponse = require('../utils/serverResponse');
 
@@ -24,12 +26,15 @@ const submitPost = async (req, res) => {
   try {
     const buffer = await formatImage(req.file, 600);
     const mediaUrl = await uploadFile(req.file.originalname, buffer);
+
     const post = {
       ...req.body,
       user: req.user.sub,
       media: mediaUrl
     };
+
     const newPost = await createPost(post);
+
     return serverResponse(res, 200, newPost);
   } catch (e) {
     return serverResponse(res, 500, {
@@ -44,6 +49,11 @@ const submitPost = async (req, res) => {
 const getPosts = async (req, res) => {
   try {
     const posts = await getAllPosts(req.query.limit, req.query.skip);
+
+    if (posts.length === 0) {
+      return serverResponse(res, 404, { message: 'No posts found' });
+    }
+
     return serverResponse(res, 200, posts);
   } catch (e) {
     return serverResponse(res, 500, {
@@ -59,15 +69,8 @@ const getPostsOfAUser = async (req, res) => {
   try {
     const posts = await getAllPostsOfUser(req.params.userInfo);
 
-    if (!posts.length) {
-      return serverResponse(res, 404, {
-        message: 'there are no posts with requested user id'
-      });
-    }
-
     return serverResponse(res, 200, posts);
   } catch (e) {
-    console.log(e);
     return serverResponse(res, 500, {
       message: 'Internal error while trying to get posts'
     });
@@ -80,13 +83,14 @@ const getPostsOfAUser = async (req, res) => {
 const getOnePost = async (req, res) => {
   try {
     const post = await getPost(req.params.postId);
+
     if (!post) {
-      return res
-        .status(404)
-        .json({ message: 'no post with requested id' })
-        .end();
+      return serverResponse(res, 404, {
+        message: "Post doesn't exist"
+      });
     }
-    res.status(200).json(post).end();
+
+    return serverResponse(res, 200, post);
   } catch (e) {
     res
       .status(500)
@@ -102,17 +106,28 @@ const deletePost = async (req, res) => {
   try {
     const post = await getPost(req.params.postId);
 
+    if (!post) {
+      return serverResponse(res, 404, {
+        message: "Post doesn't exist"
+      });
+    }
+
     if (!requesterIsAuthenticatedUser(req.user.sub, post.user)) {
       return serverResponse(res, 400, {
         message: 'Unauthorized!'
       });
     }
 
-    await removeLikesFromPost(req.params.postId);
-    await removeAllPostComments(req.params.postId);
-    await deleteFile(post.media);
-    await removePost(req.params.postId);
-    return serverResponse(res, 200, { message: 'File successfully deleted' });
+    const comments = await getCommentsOfPost(req.params.postId);
+
+    const removeLikesFromCommentPromises = comments.map(async comment => removeLikesFromComment(comment._id));
+
+    await Promise.all([removeLikesFromPost(req.params.postId),
+      removeAllPostComments(req.params.postId),
+      deleteFile(post.media),
+      removePost(req.params.postId),
+      ...removeLikesFromCommentPromises]);
+    return serverResponse(res, 200, { message: 'Post successfully deleted' });
   } catch (e) {
     return serverResponse(res, 500, {
       message: 'internal error while trying to delete a post'
@@ -126,12 +141,21 @@ const deletePost = async (req, res) => {
 const editPost = async (req, res) => {
   try {
     const postToEdit = await getPost(req.params.postId);
+
+    if (!postToEdit) {
+      return serverResponse(res, 404, {
+        message: "Post doesn't exist"
+      });
+    }
+
     if (!requesterIsAuthenticatedUser(req.user.sub, postToEdit.user)) {
       return serverResponse(res, 400, {
         message: 'Unauthorized!'
       });
     }
+
     const post = await updatePost(req.params.postId, req.body);
+
     return serverResponse(res, 200, post);
   } catch (e) {
     return serverResponse(res, 500, {
